Enabled automatic queue_handler decoration
