+++ samples/sample4.py
@@ -1,10 +1,9 @@
-# Sample - Prime Number Counter
-
-from mpcurses import queue_handler
-from mpcurses import MPcurses
-
-from time import sleep
 import logging
+from time import sleep
+
+from mpcurses import MPcurses
+from sample4_sl import get_screen_layout
+
 logger = logging.getLogger(__name__)
 
 
@@ -19,7 +18,6 @@ def is_prime(num):
         return True
 
 
-@queue_handler
 def check_primes(data, shared_data):
     range_split = data['range'].split('-')
     lower = int(range_split[0])
@@ -31,95 +29,6 @@ def check_primes(data, shared_data):
             logger.debug('prime')
         else:
             logger.debug('not prime')
-
-
-def get_screen_layout():
-    return {
-        'number_header': {
-            'position': (2, 4),
-            'text': 'Number',
-            'text_color': 14,
-        },
-        'prime_header': {
-            'position': (2, 15),
-            'text': 'Prime :',
-            'text_color': 14,
-        },
-        'not_prime_header1': {
-            'position': (1, 22),
-            'text': 'Not',
-            'text_color': 14,
-        },
-        'not_prime_header2': {
-            'position': (2, 21),
-            'text': 'Prime',
-            'text_color': 14,
-        },
-        'number': {
-            'position': (3, 1),
-            'text': '',
-            'color': 27,
-            'regex': '^checking (?P<value>\d+)/\d+$',
-            'table': True
-        },
-        'uppper_number': {
-            'position': (3, 7),
-            'text': '',
-            'color': 15,
-            'regex': '^checking \d+(?P<value>/\d+)$',
-            'table': True
-        },
-        'prime': {
-            'position': (3, 15),
-            'text': '',
-            'color': 2,
-            'keep_count': True,
-            'zfill': 5,
-            'regex': '^prime$',
-            'table': True
-        },
-        'not_prime': {
-            'position': (3, 21),
-            'text': '',
-            'color': 3,
-            'keep_count': True,
-            'regex': '^not prime$',
-            'zfill': 5,
-            'table': True
-        },
-        '_counter_': {
-            'position': (3, 27),
-            'categories': [
-                'number'
-            ],
-            'counter_text': '|',
-            'modulus': 200,
-            'color': 15,
-            # 'regex': '^total of (?P<value>\d+) numbers$',
-            'table': True
-        },
-        'total_header': {
-            'position': (14, 8),
-            'text': 'Total:',
-            'text_color': 14,
-        },
-        'prime_total': {
-            'position': (14, 15),
-            'text': '',
-            'color': 2,
-            'keep_count': True,
-            'zfill': 5,
-            'regex': '^prime$',
-        },
-        'not_prime_total': {
-            'position': (14, 21),
-            'text': '',
-            'color': 3,
-            'keep_count': True,
-            'zfill': 5,
-            'regex': '^not prime$',
-        },
-    }
 
 
 def main():
@@ -135,11 +44,11 @@ def main():
             {'range': '60001-70000'},
             {'range': '70001-80000'},
             {'range': '80001-90000'},
-            {'range': '90001-100000'},
+            {'range': '90001-100000'}
         ],
-        processes_to_start=10,
         screen_layout=get_screen_layout()).execute()
 
 
 if __name__ == '__main__':
+
     main()
